pyproject.toml
[tool.poetry]
name = "mlflow-on-aws"
version = "0.1.0"
description = ""
authors = ["anya-mb <anyamb@yandex.ru>"]
readme = "README.md"
packages = [{include = "mlflow_on_aws"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "2.84.0"
constructs = "^10.2.55, <11.0.0"
pre-commit = "^3.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# https://beta.ruff.rs/docs/configuration/
[tool.ruff]
line-length = 120
# Skip unused variable rules (`F841`).
ignore = ["F841"]
